refactor: split workspace endpoints out of server.js into controller and router file, add additional docs to s3Utils
